refactor(session): re-home the session runtime into a neutral package
The machinery that keeps a session addressable from outside its own process —
the discovery record, the authenticated loopback control socket, the
owned-session handle and the no-terminal child process — lived under `mobile/`
because the phone was the first viewer that needed it. None of it is about the
phone: `lop sessions`, `lop send` and attach all read the same records, and
wakes and background automations will be clients of the same socket. Leaving
it under `mobile/` made every non-phone consumer import a front end it does
not use.

Move `mobile/{registrant,registry,owned,child}.py` to
`local_operator/session/runtime/{server,registry,owned,process}.py`, and the
record/socket constants to `session/runtime/types.py`. `Registrant` becomes
`RuntimeServer`, aliased at both paths. Thin re-export shims stay at every old
`mobile/` module path.

The vocabulary is deliberate: a RUNTIME is the process that owns a running
session; a VIEWER is a TUI or phone attached to it. The reaper already encoded
that distinction — viewer count is not an input to whether an idle runtime
exits — and the package name now says it too.

No behaviour change. Every literal is byte-identical — `RUN_DIRNAME` stays
`"run/mobile"`, PROTOCOL_VERSION stays 5, the heartbeat intervals, the
`LOP_MOBILE_CHILD_*` spawn variables and the control thread's name are all
unchanged — because two lop binaries of different versions coexist in running
processes on one machine and must keep seeing each other's sessions.

`owned.py`'s `session_factory` import stays function-local and is now
commented as load-bearing: hoisting it would create a genuine import cycle
now that this module lives under `session/`.

Fixes a peer-send import-light guard that asserted the absence of a path
prefix rather than of the heavyweight Session graph, and which the move turned
into a false positive; it now measures the real import set in a fresh
interpreter and is proven to fail when the graph is pulled in.
